+++ R/bounceMarker.R
@@ -9,9 +9,7 @@ bounceMarkerDependency <- function() {
 }
 
 #' Add Bounce Markers to map
-#' @param map map object created by [leaflet::leaflet]
-#' @param lat numeric latitude
-#' @param lng numeric longitude
+#' @inheritParams leaflet::addMarkers
 #' @param duration integer scalar: The duration of the animation in milliseconds.
 #' @param height integer scalar: Height at which the marker is dropped.
 #' @md
@@ -22,7 +20,47 @@ bounceMarkerDependency <- function() {
 #' leaflet() %>%
 #'   addTiles() %>%
 #'   addBounceMarkers(49, 11)
-addBounceMarkers = function(map, lat, lng, duration = 1000, height = 100) {
+addBounceMarkers = function(map, lng = NULL, lat = NULL, layerId = NULL, group = NULL,
+                            icon = NULL,
+                            duration = 1000, height = 100,
+                            popup = NULL,
+                            popupOptions = NULL,
+                            label = NULL,
+                            labelOptions = NULL,
+                            options = leaflet::markerOptions(),
+                            data = leaflet::getMapData(map)) {
+
+  if (missing(labelOptions))
+    labelOptions <- leaflet::labelOptions()
+  if (!is.null(icon)) {
+    icon <- leaflet::evalFormula(list(icon), data)[[1]]
+    if (inherits(icon, "leaflet_icon_set")) {
+      icon <- leaflet:::iconSetToIcons(icon)
+    }
+    icon$iconUrl <- leaflet:::b64EncodePackedIcons(leaflet:::packStrings(icon$iconUrl))
+    icon$iconRetinaUrl <- leaflet:::b64EncodePackedIcons(leaflet:::packStrings(icon$iconRetinaUrl))
+    icon$shadowUrl <- leaflet:::b64EncodePackedIcons(leaflet:::packStrings(icon$shadowUrl))
+    icon$shadowRetinaUrl <- leaflet:::b64EncodePackedIcons(leaflet:::packStrings(icon$shadowRetinaUrl))
+    if (length(icon$iconSize) == 2) {
+      if (is.numeric(icon$iconSize[[1]]) && is.numeric(icon$iconSize[[2]])) {
+        icon$iconSize <- list(icon$iconSize)
+      }
+    }
+    icon <- leaflet::filterNULL(icon)
+  }
+
   map$dependencies <- c(map$dependencies, bounceMarkerDependency())
-  invokeMethod(map, getMapData(map), 'addBounceMarkers', lat, lng, duration, height)
+
+  if (!is.null(clusterOptions))
+    map$dependencies = c(map$dependencies, leaflet::leafletDependencies$markerCluster())
+
+  pts = leaflet::derivePoints(
+    data, lng, lat, missing(lng), missing(lat), "addBounceMarkers")
+
+  leaflet::invokeMethod(
+    map, data, "addBounceMarkers", pts$lat, pts$lng, icon, layerId, duration, height,
+    group, options, popup, popupOptions,
+    NULL, NULL, leaflet::safeLabel(label, data), labelOptions,
+    leaflet:::getCrosstalkOptions(data)
+  ) %>% leaflet::expandLimits(pts$lat, pts$lng)
 }